Convert Shortcode to Use Handlebars
Updated the code used to display our shortcodes to use Handlebars templates (as opposed to bare HTML with some PHP tags to echo out the values).

Created a basic TemplateManager class, with an eye towards storing our compiled templates in the database. For the time being, and until we test the validity of storing templates that way, I'm reading them from the file system. This method will allow for people to keep their current FTP based workflow, and allow the files on disk to be the canonical templates while we do more development.